requirements.txt
gspread==5.12.4
google-auth==2.3.3
pytest==8.3.3
pytest-mock==3.10.0